+++ modules/cooldownChecker.js
@@ -1,5 +1,5 @@
-function getIDByType(bot, message, command) {
-	let cdType = bot.commands.get(command).cooldown.type;
+function getIDByType(bot, message, commandName) {
+	let cdType = bot.commands.get(commandName).cooldown.type;
 	if (cdType == "user") {
 		return message.author.id;
 	} else if (cdType == "channel") {
@@ -11,30 +11,29 @@ function getIDByType(bot, message, command) {
 	}
 }
 
-function findCooldown(bot, id, command) {
-	return bot.cache.recentCommands.find(cd => cd.id == id && cd.command == command);
+function findCooldown(bot, id, commandName) {
+	return bot.cache.recentCommands.find(cd => cd.id == id && cd.command == commandName);
 }
 
-function addCooldown(bot, message, command) {
-	let cdID = getIDByType(bot, message, command);
-	let cdTime = bot.commands.get(command).cooldown.time;
+function addCooldown(bot, message, commandName) {
+	let cdID = getIDByType(bot, message, commandName);
+	let cdTime = bot.commands.get(commandName).cooldown.time;
 	bot.cache.recentCommands.push({
 		id: cdID,
-		command: command,
+		command: commandName,
 		resets: Number(new Date()) + cdTime,
 		notified: false
 	})
-	setTimeout(removeCooldown, cdTime, bot, cdID, command);
+	setTimeout(removeCooldown, cdTime, bot, cdID, commandName);
 }
 
-function removeCooldown(bot, id, command) {
-	bot.cache.recentCommands.splice(findCooldown(bot, id, command), 1);
+function removeCooldown(bot, id, commandName) {
+	bot.cache.recentCommands.splice(findCooldown(bot, id, commandName), 1);
 }
 
 module.exports = {
 	check: (bot, message, command) => {
-		let cdType = bot.commands.get(command).cooldown.type;
-		let checkedCd = findCooldown(bot, getIDByType(bot, message, command), command);
+		let checkedCd = findCooldown(bot, getIDByType(bot, message, command.name), command.name);
 		if (checkedCd) {
 			if (!checkedCd.notified) {
 				checkedCd.notified = true;
@@ -48,9 +47,9 @@ module.exports = {
 				let toSend = `⛔ **Cooldown:**\n*${cdMessages[Math.floor(Math.random() * cdMessages.length)]}*` + "\n" +
 				`This command cannot be used again for **${((checkedCd.resets - Number(new Date())) / 1000).toFixed(1)} seconds**`
 				if (message.guild) {
-					if (cdType == "channel") {
+					if (command.cooldown.type == "channel") {
 						toSend += " in this channel";
-					} else if (cdType == "guild") {
+					} else if (command.cooldown.type == "guild") {
 						toSend += " in this server";
 					}
 				}