+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 # 1.X (current)
 
+## 1.3.4
+
+- Added more image commands
+
+## 1.3.0
+
+- Added a new category Text Editing
+
 ## 1.2.26
 
 - Move commands into category files; `commands` now has no subfolders.